docs/examples/simple/requirements.txt
# Simple A2A Agent Dependencies
phlow>=0.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
google-genai>=0.1.0
requests>=2.31.0
